Facilitando a injeção com RequiredArgsConstructor

+++ 04- API'sRestful/src/main/java/io/cesa/libraryAPI/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.cesa.libraryAPI.service;
 
+import io.cesa.libraryAPI.exceptions.OperacaoNaoPermitidaException;
 import io.cesa.libraryAPI.model.Autor;
 import io.cesa.libraryAPI.repository.AutorRepository;
+import io.cesa.libraryAPI.repository.LivroRepository;
+import io.cesa.libraryAPI.validator.AutorValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,20 +17,41 @@ public class AutorService {
 
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
+    public AutorService(AutorRepository repository,
+                        AutorValidator validator,
+                        LivroRepository livroRepository){
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return  repository.save(autor);
     }
 
+    public void atualizar(Autor autor){
+         if(autor.getId() == null){
+             throw new IllegalArgumentException("Para atualizar, é necessario que o autor já  esteja salvo na base ");
+         }
+         validator.validar(autor);
+         repository.save(autor);
+    }
+
+
     public Optional<Autor> obterPorId(UUID id){
+
         return repository.findById(id);
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)){
+            throw  new OperacaoNaoPermitidaException(
+                "Não é permitido excluir um Autor possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -47,5 +71,8 @@ public class AutorService {
          return  repository.findAll();
     }
 
+    public boolean possuiLivro(Autor autor){
+        return  livroRepository.existsByAutor(autor);
+    }
 
 }
